+++ TODO.md
@@ -1,6 +1,6 @@
 # TODO
 
-- [ ] [WORKFLOW] Add tests. Becoming a nightmare without.
+- [ ] [WORKFLOW] Add tests.
 - [ ] [IDEA] Other competition formats (UCL?).
 - [ ] [APPEARANCE] Dash: DataTable filtering, more info, DAQ widgets, etc.
 - [ ] [WORKFLOW] Incorporate a workflow.
