+++ utils/snippets.rst
@@ -12,15 +12,6 @@ Python commands
 
 - Print HDF5 keys: ``python -c "import pandas as pd;store=pd.HDFStore('fbd_storage.h5');print(store.keys())"``
 
-
-
-- First item
-- Second item
-
-  - Sub item
-
-- Third item
-
 =========== ========
 Country     Capital
 =========== ========
