--- version-info.json
@@ -1,9 +1,0 @@
-{
-    "name": "elo_football_ratings",
-    "version": "0.7.0",
-    "author": {
-        "name": "Alex Munger",
-        "email": "munger.alex@gmail.com",
-        "url": "https://amunger3.github.io"
-    }
-}
